rename files, set auto_paginate to False
