+++ parcllabs/__init__.py
@@ -8,7 +8,7 @@ api_key: Optional[str] = None
 api_base = DEFAULT_API_BASE
 
 
-from parcllabs.plabs_client import ParclLabsClient
+from parcllabs.parcllabs_client import ParclLabsClient
 
 
 from parcllabs.services.investor_metrics import (
