+++ parcllabs/services/rental_market_metrics.py
@@ -1,6 +1,6 @@
 from typing import Any, Mapping, Optional, List
 
-from parcllabs.services.parcl_labs_service import ParclLabsService
+from parcllabs.services.parcllabs_service import ParclLabsService
 
 
 class RentalMarketMetricsGrossYield(ParclLabsService):
